Win64 compilation.
git-svn-id: http://www.bci2000.org/svn/trunk@3962 7ef59cdf-0624-0410-a474-e34ce222aafa

+++ src/contrib/ExternalLinks/BCI2000Command/BCI2000Command.cpp
@@ -72,7 +72,7 @@
 { \
   const char* pValue = NULL, \
             * pOption = OPTION( name ); \
-  int len = ::strlen( pOption ); \
+  size_t len = ::strlen( pOption ); \
   int i = 1; \
   while( !pValue && i < argc && *argv[i] == '-' ) \
   { \
